+++ packages/file_selector/file_selector_ios/CHANGELOG.md
@@ -1,5 +1,6 @@
-## NEXT
+## 0.5.4
 
+* Adds `canCreateDirectories` parameter to `FileDialogOptions` to control the visibility of the New Folder button in file dialogs on supported platforms.
 * Updates minimum supported SDK version to Flutter 3.29/Dart 3.7.
 
 ## 0.5.3+2